examples/evaluation/requirements.txt
matplotlib
python-dotenv
azure-ai-projects
azure-ai-evaluation